requirements.txt
streamlit
openai
PyGithub
python-dotenv